Redirect ServerFoundCallbacks to main thread
Executing ServerFoundCallback on the worker thread causes its linked delegates to also be executed on the worker thread. If the delegate calls ANYTHING related to Unity engine, this either causes an error or your calls to be completely ignored. This commit changes this, and invokes the callback inside the Update() function, which ensures that the callback occurs in the main thread.