Continue server discovery even after one server is found
Previously, the server discovery would instantly stop after a server was found, so if there were more than one server on your network, you were out of luck. This fixes that, and allows the script to discover more than the first server by saving the previously found servers on a list, and then comparing the newly found endpoints to the previous ones to check if they are newly discovered servers.